add confirmation prompt for cluster uninstallation

+++ cli/main.go
@@ -1,9 +1,13 @@
 package main
 
 import (
+	"bufio"
+	"fmt"
 	"geet.svck.dev/urumo/k3sd/cluster"
 	"geet.svck.dev/urumo/k3sd/utils"
 	"log"
+	"os"
+	"strings"
 )
 
 func main() {
@@ -18,11 +22,23 @@ func main() {
 	if err != nil {
 		log.Fatalf("failed to load clusters: %v", err)
 	}
+
 	if utils.Uninstall {
-		cluster.UninstallCluster(clusters, logger)
+		reader := bufio.NewReader(os.Stdin)
+		fmt.Print("Are you sure you want to uninstall the clusters? (yes/no): ")
+		response, _ := reader.ReadString('\n')
+		response = strings.TrimSpace(strings.ToLower(response))
+
+		if response == "yes" {
+			cluster.UninstallCluster(clusters, logger)
+		} else {
+			fmt.Println("Uninstallation canceled.")
+			return
+		}
 	} else {
 		cluster.CreateCluster(clusters, logger, []string{})
 	}
+
 	if err := cluster.SaveClusters(utils.ConfigPath, clusters); err != nil {
 		log.Fatalf("failed to save clusters: %v", err)
 	}
